Punish update
Every Offence is RPT logged with UUID.

Admins are except (but still logged), as locking out the admin's controls is not a good idea. And sometimes immediate action is nessary.

Civs no longer exec TK punish, it was a little too harsh.

Punishment.sqf has had a major rework, it's easy looking at the code.