Adds Breaching Charges for Vehicles
Adds a small feature, that lets engineers breach open armored vehicles with explosives. Duration, needed explosive and damage dealt to vehicle depends on the vehicle. APCs need 40 sec, one explosive charge and get a damage of max 30%. Tanks need 60 sec, one explosive satchel and get a damage of max 50%. Cancelling the action cost nothings and deals no damage. If the vehicle is already damaged and the breach operation exceeds 100% damage, the vehicle will blow up with a combined explosion of the explosive.

On success, the explosion sound will play and the remaining crew will have a chance to die. All surviving crew member will exit and surrender (Dont panik and shoot them). The vehicle will be unlocked at this point.