Fixed flag exploit by adding safety check (#419)
* Fixed #409 by adding safety check

I added a isGettingCaptured variable to the flag, which is getting set to true once someone is capturing the flag. Every further call of the call will abort and hint the player about it.

* Fixed bad case, put reset to markerChange

Spoffy found a way to exit the flag without reseting the isGettingCapture variable, which I  missed due to the formating. Added a reset for this case

Pushed the last reset into markerChange, when the new functions are added too, to avoid having a small window in which the exploit could have worked

* Fixed it, this time for real

Changed the system to lock the flag for 30 seconds, hopefully avoiding the issue of the flag possibly being locked forever.