Fixed wrong variable names
This may lead to another problem where the convoy gets assigned a ID (lets say 500) arrives on target and the ID 500 gets reseted after 10 minutes. The code to delete it after 30 is still sleeping for 20 minutes.

If the convoy creation code manages to create another convoy with ID 500 in this 20 minutes, the number will be reseted after 20 minutes.

Not sure if that creates any problems besides making logs a bit harder to read (it should not) but the system may needs a complete rework.