Prevents invalid items from RHS/3CB/IFA making it into crates
Adds picture filter to crates
We remove anything without a picture, because it's a surprisingly good indicator if whether something
is actually a valid item or not.
Despite all the filtering, we still get a few RHS guns, etc that are for APCs, but are still classed the item type as normal weapons.
This is a pretty hard filter that removes anything that shouldn't be in there - I'm hoping it isn't prone to false positives!