Added function to clean up deleted marker from convoy array
This fixes a bug which would try to select deleted markers for the intel reveal and therefor doesn't show anything. It should also stop the array from getting too large with unused data.

Fixed a small typo in selectIntel which retrieved the occupants convoy markers for the selection instead of the invaders ones