Fix exploit where crates could be loaded into a vehicle while undercover (#974)
Loot crates could be loaded into vehicles without breaking undercover, so it was straightforward to farm them in the early game by driving in with civvie clothes and an offroad/truck. This PR breaks undercover when a JNL crate is loaded.

The local setCaptive should be fine here, as the player is always local to an addAction.