Wurzel aggro rework (#975)
Changed the aggression system as it was just a number going from 5 to 100 without much players can do about it.

As the idea is to change the aggression value to something punishing aggressive behaviour and rewarding slow stealthy attacks, I implemented the aggression value as a stack, carrying all aggro changes at once. This stack is updated every minute and each aggro change reduces until i drops to 0 and gets deleted. 

The general number from 0 to 100 stays, but it hidden behind words. Players only get shown that the aggro level is one of "Low", "Medium", "High", "Very high" or "Extreme". Each of them covering roughly 20 points. Roughly as it takes a few more points to switch to another level, "Medium" for example covers 17.5 - 42.5, while "Low" covers 0 - 22.5.

The stack looks like this: 
`[[+ 15, -0.5], [- 3, +0.28]]`
The first number is the actual aggression change, the second as the decay per minute. The decay per minute is decided by the current aggression in the moment, when the new aggression is added.