Split patrolCA into decision & execution functions (#1200)
A synchronization bug in patrolCA was causing multiple patrolCAs to be sent to the same location. It's an old bug but was far more obvious with six vehicles per patrolCA.

This was resolved by splitting patrolCA into decision and execution functions. The decision part (still called patrolCA) is server-only and strictly fenced, so it won't start examining a second patrolCA request until the previous one has been decided. This function calls createQRF to handle the actual spawning and management of the QRFs, or launches airstrikes itself.

The underlying logic is mostly unchanged, except for some bugs discovered although the way and some mostly-uncontroversial balance changes:

- Halved the vehicle count of position-target patrolCAs (they're now 1-3 vehicles. Used to be 1-6).
- Stop sending vehicles once maxUnits is reached.
- Increase post-termination delay before the next patrolCA on the same target.
- Add some minor vehicle adjustment for air/tank targets.
- Removed ability for cargoSeats to send more than one squad per vehicle.
- Fixed an issue where air-target-type patrolCAs would never terminate.
- Force air vehicles to return to base on termination.
- Removed some old code that was creating rebel squads after a capture.
- Fixed a bug where smallCAmrk was incorrectly preserved on save/load.

I would consider this a temporary patch to get things into a workable state for 2.3. The whole QRF/support system needs to be ripped out and replaced after that.