Group leader now gets passed when the SL is not longer capable (#1262)
### What have you changed and why?
Information:
The squad leader is important for AI groups as he does most of the calculations and is also the one calculating the path and the waypoints. When he dies, the engine passes command to the second unit. However Antistasi has to cases in which the leader does not die, but still get incapable of executing commands, these being surrendering or going uncon.

In that two cases, the group stops to move and gets more or less useless until the SL dies. To avoid that behaviour, the squad lead is now passed onto the next active unit in the group, keeping the group active until no AI left in fighting state.

For the support rework: If the SL calls for help and dies, then passing command to the next one, the call for support is aborted, as the new squadleader is different which is checked. The new squad leader can instantly call another support, unless the group called support in the last 10 minutes.

### Please specify which Issue this PR Resolves.
closes #1173 

### How can the changes be tested?
Steps: 
Spawn in an outposts by teleporting there.
Check in the zeus interface which unit is a squadleader (static gunners are best)
Shoot the unit uncon
Check in zeus again and see that the SL has been passed on