Classify ace gadgets and add them to loot crates (#1440)
## What type of PR is this.
1. [ ] Bug
2. [ ] Change
3. [X] Enhancement

### What have you changed and why?
Created a category for ACE electronic gadgets (none of the previous categories fitted) and added the Kestrel, AtragMX and HuntIR to it. The HuntIR was already force-added to lootItem in the ACE compat template, so that was removed to avoid double entries.

### Please specify which Issue this PR Resolves.
closes #1392 

### Is further testing or are further changes required?
1. [X] No
2. [ ] Yes (Please provide further detail below.)

May want to increase the number of different items dropped in crates for a better chance of finding this stuff. There are around 40 different classes in lootItem in ACE/RHS, and a maximum of four different items by default.

### How can the changes be tested?
Check what's in the following vars:
```
allGadgets
unlockedGadgets
allItems
unlockedItems
lootItem
```

Addition:
Added ACE_microDAGR and ACE_DAGR (Bob)


Co-authored-by: Bob-Murphy <50651578+Bob-Murphy@users.noreply.github.com>