Restrict loot crate respawns (#1494)
## What type of PR is this.
1. [ ] Bug
2. [ ] Change
3. [X] Enhancement

### What have you changed and why?
Loot crates no longer respawn every time an outpost/seaport/airport garrison is spawned. Instead, stealing or destroying a crate adds a cooldown variable to that garrison, which is decreased on ticks only if the garrison is full. Hence a garrison that isn't being reinforced will not respawn loot crates. Locations with larger garrisons have shorter cooldowns, typically an hour for a large outpost, two hours for a small one or half an hour for an airfield.

This prevents farming a near-empty outpost that's reinforcement-blocked by killzones, and reduces (but doesn't prevent) exploiting outposts that are relatively easy to steal crates from for geographical reasons. It's also logical, because outposts don't send out patrols (to confiscate loot, or whatever the hell they're doing to get those weapons) unless they're at capacity.

### Please specify which Issue this PR Resolves.
closes #1446

### Is further testing or are further changes required?
1. [X] No
2. [ ] Yes (Please provide further detail below.)

### How can the changes be tested?
Code to return garrison capacity, real garrison count (can be larger) and crate cooldown for a garrison:

```
_marker = "outpost_25";
[[_marker] call A3A_fnc_garrisonSize, count (garrison getVariable _marker), garrison getVariable [_marker+"_lootCD", 0]]
```


* Add a system to block loot crates from respawning until reinforced.
- Locations need to maintain a full garrison for ~30-120 minutes.
- Locations with larger garrisons regenerate their loot crate more quickly.
- Saved as part of garrison data, backwards compatible.

* Fix ammobox destruction case for loot cooldowns and cleanup