New jnl (#1588)
## What type of PR is this.
1. [ ] Bug
2. [ ] Change
3. [x] Enhancement

### What have you changed and why?
Information: JNL replacement Logistics (aka newJNL)
JNL was full of bugs from lack of maintenance and extremely un-flexible, and with the common agreement of moving away from all the JN systems, there was no need to keep JNL in favor of another system as it no longer would have integration in a partially functional system anymore.

So with this in mind and a desire for a more flexible logistics system; voiced by the community during testing of LTC, I decided to resolve this, as JNL was riddled with bugs and I needed to completely replace the core functionality; not necessarily heavily modify the surrounding logic. I decided to fully replace JNL with a new system.

The new system will be consisting of vehicles with x number of EQUAL nodes to handle both cargo and weapon loading; these node will be equal in size, the cargo would then “use” x amount of these nodes depending on the defined size of the cargo, from the smallest “surrender crates” with a size of 1 to the large; and realistically unloadable in with the current defined vehicles if matching vehicles to mod set “ZU-23” with a size of 7 nodes.
Per request the new system features a blacklist for the weapons defined to block it from being mounted/loaded onto certain vehicles(models) for example mortars on assault boats, loading weapons in a covered transport vehicle or in the back of a closed van; this is done with a general weapon loading ban on covered/closed vehicles (manually defined).

To summarize JNL is unmaintained, buggy, and inflexible; to the point where a new system was the best option.

### Please specify which Issue this PR Resolves.
closes #1652 

### Please verify the following and ensure all checks are completed.

1. [x] Have you loaded the mission in singleplayer?
2. [x] Have you loaded the mission in LAN host?
3. [x] Have you loaded the mission on a dedicated server?

### Is further testing or are further changes required?
1. [x] No
2. [ ] Yes (Please provide further detail below.)

### How can the changes be tested?
Steps: preferable dedicated server, use all available options of the system and generally try to break it, its only use will be trough the action added with ``A3A_fnc_logistics_addLoadAction;`` 

********************************************************
Notes: when we get around to doing MIE the only change will be how the nodeInit is done, as then it will be built in templates, then distributed by nodeInit from the server only, in contrast with now where they are run on each client locally.


* integrated new JNL, old JNL disconnected, files left alone

striped out old "Steal static" action as well as it was obsolete...

* combined weapons and blacklist arrays

* moved the valid static check to addLoadAction

* added support for classname based static list

gets the model when inits the nodes based of the classname

* RHS nodes added, finished converting to class-named definitions

weapons and blacklist not fully updated

* added 3cb to init, finished the weapon list and coveredVehicles

next step integration with MIE system meer is working on

* reverted server execution

for some reason the attachements offset list couldnt find the modle for some items, benefits of server execution is not grate enough in my opinion for it to be wort the time and effort to find and resolve the issues with moving it to server side execution

* updated get in weapon action

to better handle breaks in removal of action code

* added locality switch before blocking seats

pesky lock seats commands needing the vehicle to be local to the executing machine

* fix for the broken seat locking on DS

* second go at locked cargo fix for DS (tested and working)

* updated origin branch with core changes in MIEintegration

* fixed a cupple of missed JN load calls

salvage rope needs added functionality

* salvage fix

* added rough header, this will all change when we merge MIE

as it will be run only server side. Note weapon EH may be spammed multiple times, check eh globallity

* added the new city supplies crate to cargo attachment offset

* fixed wrong indentation

* replaced old helper tool and guide

* changed to dinfIf in addLoadAction

* switched to customHint

* tabs to spaces (indentation) and functions update for the helper tool

forgot to include it in that commit

* correcter an error in helper tool boundary render

* made the helper tool maximize the space

* fixed bad invalid weapon detection in logistics_addLoadAction

and fixed fixed bad type check

* fixed param of call being passed as object and not array with object

* moved to bounding box relative dismount point

Y offset based on weapon attachment offset from vehicle model center.

* split can load in two functions

canLoad - checks if vehicle can load cargo, retruns error code or array to be passed to load function

tryLoad- handles standard loading procedure with feedback for player

* fix bad rotation on multiplayer

* correct load calls for salvage rope

* fixed typo in addWeaponAction

* added podnos and m252, fixed tabs indent

* brought JNL up to new standard

* corrected typos

* added public node count function ++

* addapted for pr 1481, vehCargoTrucks

* fixed logical error in addOrRemoveObjectMass

* bit of childproofing

* requested changes: calebSerafin

* requested changes pt2: calebSerafin

* added A3A prefix to logistics global data

* logic change request: calebSerafin

* travis being funny

* dont be mean travis

* i dont even know anymore travis