Testing light (#2040)
## What type of PR is this.
1. [ ] Bug
2. [ ] Change
3. [x] Enhancement

### What have you changed and why?
Information:
Adds a purchasable light source to the vehicle box. 
Allowing for a brighter HQ during the night.
Allows for seeing the map better at night during early game missions.
If someone wants to delete the light the just garage it and gets the refund.


    
Added a vehiclelightsource variable to the rebel templates, allowing for the change to a different 
light source object if players request a different source later.

Added a event handler to rotate the light. Rotating the light allows for exactly placement of the light rays.

players can turn off the changes if they disable parameter.
### Please specify which Issue this PR Resolves.
closes #1620 

### Please verify the following and ensure all checks are completed.

1. [x] Have you loaded the mission in LAN host?
2. [ ] Have you loaded the mission on a dedicated server?

### Is further testing or are further changes required?
1. [x] No
2. [ ] Yes (Please provide further detail below.)

### How can the changes be tested?
Steps: 


********************************************************
Notes:

The rotate item function is not upvector safe. 

This pull request is intended to make playing in the night easier for the early game when night vision is unavailable.
An example of the QOL improvement would be looking at the map at night. Most of the item I stand in front of a vehicle with its lights on, but with this purchasable light I can stand anywhere.  

might need to be tested with two people garaging the light.

* added a UtilityItems, a light function and more

* changed buying lights

* pretemplates (working)

* after template (working)

* added params

* RM template, Server Param, changed array

* changed by review

* second review

* ending checks

* spelling fixes

* spacing fixed