Quick player-count balance for QRF/singleAttack/wavedCA (#2124)
### What type of PR is this.
1. [ ] Bug
2. [X] Change
3. [ ] Enhancement

### What have you changed and why?
Softer version of #2079 based on the principle that aggro is already somewhat player-count dependent, so it should be used additively rather than multiplicatively with direct player count balancing. Uses a bit of randomness to smooth out the boundaries. Typical vehicle counts:

QRF:
1 player, 25 aggro: 1.5 vehicles.
9 players, 50 aggro: 2.5 vehicles.
26 players, 100 aggro: 3.5 vehicles.

singleAttack (typically higher aggro because a marker was just captured):
1 player, 50 aggro: 3 vehicles.
9 players, 75 aggro: 4.8 vehicles.
26 players, 100 aggro: 7.5 vehicles.

wavedCA (+2 vehicles for first wave)
1 player: 5.5 vehicles first wave, 3.5 subsequent
9 players: 7 vehicles first wave, 5 subsequent
26 players: 10 vehicles first wave, 8 subsequent

wavedCA isn't aggro-dependent because it's kinda odd to increase the attack size because you just took a marker elsewhere. It's a bit flatter than the others because the garrisons + statics are more useful.

### Please specify which Issue this PR Resolves.
Supersedes the other half of #2079

### Please verify the following and ensure all checks are completed.
1. [X] Have you loaded the mission in LAN host?
2. [ ] Have you loaded the mission on a dedicated server?

### Is further testing or are further changes required?
1. [X] No
2. [ ] Yes (Please provide further detail below.)

Notes:
* Quick player-count balance for QRF/singleAttack/wavedCA

* Improve logging and occ vs inv behaviour

* Move playerscale out into a tiny function