Move A3A_climate init to initVarCommon for sanity and HC-functionality (#2352)
A3A_climate was declared in selector.sqf (which only runs on the server) and not published, but used in combatLanding which runs on HCs. This probably caused singleAttack combat landings to follow the wrong path with HCs present.

selector.sqf seemed like a terrible place for it anyway, so I moved A3A_climate to initVarCommon which runs early and everywhere.